added stats for the major aplicon classifications
