+++ caper/templates/pages/admin_stats.html
@@ -97,13 +97,38 @@
                 <tbody>
                     <tr>
                         <td>Projects</td>
-                        <td>{{ site_stats.public_proj_count }}</td>
-                        <td>{{ site_stats.all_private_proj_count }}</td>
+                        <td>{{ site_stats.public_proj_count|default:"SITE STATS REGENERATION NEEDED" }}</td>
+                        <td>{{ site_stats.all_private_proj_count|default:0 }}</td>
                     </tr>
                     <tr>
                         <td>Samples</td>
-                        <td>{{ site_stats.public_sample_count }}</td>
-                        <td>{{ site_stats.all_private_sample_count }}</td>
+                        <td>{{ site_stats.public_sample_count|default:0 }}</td>
+                        <td>{{ site_stats.all_private_sample_count|default:0 }}</td>
+                    </tr>
+                    <tr>
+                        <td>ecDNA</td>
+                        <td>{{ site_stats.public_amplicon_classifications_count.ecDNA }}</td>
+                        <td>{{ site_stats.all_private_amplicon_classifications_count.ecDNA }}</td>
+                    </tr>
+                    <tr>
+                        <td>BFB</td>
+                        <td>{{ site_stats.public_amplicon_classifications_count.BFB }}</td>
+                        <td>{{ site_stats.all_private_amplicon_classifications_count.BFB }}</td>
+                    </tr>
+                    <tr>
+                        <td>Complex non-cyclic</td>
+                        <td>{{ site_stats.public_amplicon_classifications_count.Complex_non_cyclic|default:0 }}</td>
+                        <td>{{ ssite_stats.all_private_amplicon_classifications_count.Complex_non_cyclic|default:0 }}</td>
+                    </tr>
+                    <tr>
+                        <td>Linear amplification</td>
+                        <td>{{ site_stats.public_amplicon_classifications_count.Linear_amplification|default:0  }}</td>
+                        <td>{{ site_stats.all_private_amplicon_classifications_count.Linear_amplification|default:0 }}</td>
+                    </tr>
+                    <tr>
+                        <td>unclassified</td>
+                        <td>{{ site_stats.public_amplicon_classifications_count.Unclassified|default:0 }}</td>
+                        <td>{{ site_stats.all_private_amplicon_classifications_count.Unclassified|default:0 }}</td>
                     </tr>
                 </tbody>
             </table>
